requirements.txt
flask
numpy
pandas
scikit-learn
joblib
seaborn
matplotlib
plotly
folium
requests
Pillow
python-dotenv
flask-sqlalchemy
flask-login
flask-wtf
geopy
google-generativeai
mysql-connector-python
werkzeug